+++ modules/energy/bending_tilt_out.py
@@ -41,6 +41,7 @@ def compute_energy_and_gradient_array(
         tilt_grad_arr=tilt_out_grad_arr,
         kappa_key="bending_modulus_out",
         cache_tag="out",
+        div_sign=1.0,
     )
 
 
@@ -63,6 +64,7 @@ def compute_energy_and_gradient(
         tilts=tilts_out,
         kappa_key="bending_modulus_out",
         cache_tag="out",
+        div_sign=1.0,
         compute_gradient=compute_gradient,
     )